fix(plugin): Message router was not disposed
Everytime a new ExcalidrawWebView was created, a new
CefMessageRouter was registered, but never removed,
once the web view is closed.
As such when a new Webview was instantiated it used the
old message router in front which used the old associated
lifetime which was not alive. This prevented to triggering
the `IPropertyView.advise` lambda.

Moreover, the handler itself returns true, which indicates
the request will not trickle down to next handlers.

This is now fixed by properly removing and disposing of
the message handler.