fix(plugin): Fix IllegalStateException when initializing editor
When an ExcalidrawEditor is created, IntelliJ may perform
a `ActionUtil.performDumbAwareUpdate` before the editor is
ready, selected, or the attached to project.
This behavior triggers a safeguard in the plugin, the ISE.

Instead, this change makes the action code more lenient if an
editor is not found.

```
java.lang.IllegalStateException: No excalidraw editor found
	at com.github.bric3.excalidraw.actions.ToggleSaveOptionAction.getSaveOptions(ToggleSaveOptionAction.kt:18)
	at com.github.bric3.excalidraw.actions.ToggleWatermarkSaveOptionAction.isSelected(ToggleWatermarkSaveOptionAction.kt:8)
	at com.intellij.openapi.actionSystem.ToggleAction.update(ToggleAction.java:72)
	at com.intellij.openapi.actionSystem.ex.ActionUtil.performDumbAwareUpdate(ActionUtil.java:178)
```