refactor(plugin): Change custom auto-save to IDE based cadence
Previous auto-save mechanism kicked in everytime the scene was
modified, and it triggered a disk write. This is causing too
many writes, moreover with images it could cause higher CPU
usage. Instead, this refactoring relies on IntelliJ's
auto-save.

The file is saved when a Save All event is triggered, ie the
IJ's auto-save. And when the editor is deselected, however
there is currently one issue when closing the editor, because
the JCEF component is disposed right after `deselectNotify`
which prevents the Excalidraw app to send back the payload.

Also, Cmd+S is not handled by IntelliJ but by JCEF.

Side-note: using `let` prevent reuse of JS variable in
`ExcalidrawWebViewController`.