fix: Saving files
When editor was closed, the IntelliJ dispose mechanism was
called before the messaging system had a chance to communicate
with the JCEF view and get the payload back.

This commit _phases_ the save and dispose mechanisms in order to wait
for the save action to be performed before disposing.

This fix also hopes to properly register disposable components
to prevent ObjectTree stacktraces.

Fixes: #20