fix(assets): Upgrade gulp to modules to support newer yarn, and co
Upgrading react-script to 5.0.1 broke gulp script, maybe because
the 'del' package is not anymore shipped. Adding it to the `package.json`.
fixed the missing module but broke the gulp file because the 'del'
moved to pure ESM (ES Modules = ECMAScript Modules), which means the `gulpfile.js`
needs refactoring to a module in order to keep using 'del'.