feature: Load embedded images from excalidraw files
This was contributed by @lionelhorn. Approach is to wait for
the "update" message that has the embedded files then create
`initialData` with the files and only then render the
Excalidraw component.

Mobx is used for state management replacing react `useState`.
Which is easier to get the code a bit cleaner code
than passing the setters of `useState` to the bridge.

Co-authored-by: Lionel <dev@lionelhorn.com>